refractored views - added post view
