+++ Surrility/EditorViewController.swift
@@ -7,7 +7,6 @@
 //
 
 import UIKit
-import SceneKit
 import AVFoundation
 import Photos
 import Firebase
@@ -17,7 +16,7 @@ class EditorViewController: UIViewController {
     //UI OUtlets
     @IBOutlet weak var SliderA: UISlider!
     @IBOutlet weak var SliderB: UISlider!
-    @IBOutlet weak var progressView: UIProgressView!
+    //@IBOutlet weak var progressView: UIProgressView!
     @IBOutlet weak var picView: UIImageView!
     @IBOutlet weak var extractButton: UIButton!
     @IBOutlet weak var UISelector: UISegmentedControl!
@@ -26,6 +25,8 @@ class EditorViewController: UIViewController {
     var capturedPhoto: AVCapturePhoto?
     
     //internal variables
+    var downSampledImage: UIImage?
+    var frame: CloudFrame?
     var depthDataMap: CVPixelBuffer?
     var colorDataMap: CVPixelBuffer?
     var depthFilter: DepthImageFilters?
@@ -63,10 +64,19 @@ class EditorViewController: UIViewController {
                     return
             }
             
+            self.frame = frame  //store the frame for segue
+            
             //add the frame to the cloud
             //self.addPCMJSON(frame: frame)
-            self.addPCM(frame: frame)
+            //self.addPCM(frame: frame)
         }
+    }
+    
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        let DestinationViewController : PostViewController = segue.destination as! PostViewController
+        
+        DestinationViewController.ImagePreviewPlane.image = downSampledImage
+        DestinationViewController.frame = self.frame
     }
     
     @IBAction func UISelectedChanged(_ sender: Any) {
@@ -142,17 +152,6 @@ class EditorViewController: UIViewController {
         super.didReceiveMemoryWarning()
         // Dispose of any resources that can be recreated.
     }
-    
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destinationViewController.
-        // Pass the selected object to the new view controller.
-    }
-    */
 
 }
 extension EditorViewController {
@@ -197,7 +196,7 @@ extension EditorViewController {
         filter.setValue(1.0, forKey: "inputAspectRatio")
         let outputImage = filter.value(forKey: "outputImage") as! CIImage
         
-        //let outputImage = filter.value(forKey: "outputImage") as! CIImage
+        self.downSampledImage = UIImage(ciImage: outputImage) //store for later use in segue
         
         guard let colorBufferImage = tools.convertCIImageToCGImage(inputImage: outputImage) else {
             return nil
@@ -253,93 +252,14 @@ extension EditorViewController {
         DispatchQueue.main.async {
             self.SliderA.isHidden = status ? false:true
             self.SliderB.isHidden = status ? false:true
-            self.progressView.isHidden = status ? true:false;
+            //self.progressView.isHidden = status ? true:false;
         }
-    }
-    
-    func updateUserRecord(downloadURL: String){
-        //create a referene to the database
-        self.ref = Database.database().reference()
-        
-        //get the serverTimeStamp
-        let timeStamp = Firebase.ServerValue.timestamp()
-        
-        //update our records to include the user's picture
-        let uid: String = UserId!
-        let tempRef = self.ref?.child("Posts").childByAutoId().childByAutoId().parent
-        //add download url
-        tempRef?.child("Path").setValue(downloadURL)
-        //add who posted me
-        tempRef?.child("UID").setValue(uid)
-        //setup likes
-        tempRef?.child("likes").setValue(0)
-        //add timestamp
-        tempRef?.child("TimeStamp").setValue(timeStamp)
     }
     
     func moveAction(){
         //make sure this is alwas performed on the UI thread otherwise we'll get rando results
         DispatchQueue.main.async {
             self.navigationController?.popViewController(animated: true)
-        }
-    }
-    
-    func addPCM(frame: CloudFrame){
-        
-        //create a storage reference from our storage service
-        let storageRef = storage.reference()
-        
-        //Create a child reference
-        //ImagesRef now points to "images"
-        let pcmStorageRef = storageRef.child("pcms").child(UserId!)
-        
-        let dateformatter = DateFormatter()
-        dateformatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSSZ"
-        dateformatter.timeZone = NSTimeZone(name: "UTC")! as TimeZone
-        let now = dateformatter.string(from: Date())
-        let fileName = now + ".pcm"
-    
-        let spaceRef = pcmStorageRef.child(fileName)
-        let uploadMetaData = StorageMetadata()
-        uploadMetaData.contentType = "image/pcm"
-        
-        //convert cloudframe to data
-        let encoder = JSONEncoder()
-        let data = try! encoder.encode(frame)
-        
-        //let fstring: String = String(data: data, encoding: String.Encoding.utf8)!
-        //let pcmString = fstring.toBase64()
-        
-        let uploadTask = spaceRef.putData(data, metadata: uploadMetaData) { (metadata, error) in
-            if(error != nil) {
-                print("ERROR BILL ROBINSON \(String(describing: error))")
-            } else {
-                print ("Upload complete! \(String(describing: metadata))")
-                
-                // Fetch the download URL
-                storageRef.child("pcms").child(UserId!).child(fileName).downloadURL(completion: { (FileName_url, error) in
-                    if (error != nil) {
-                        print("Error getting Download Url")
-                        return
-                    } else {
-                        // Get the download URL for 'images/stars.jpg'
-                        print(FileName_url?.absoluteString)
-                        //put the download url for the record in the real-time database
-                        self.updateUserRecord(downloadURL: (FileName_url?.absoluteString)!)
-                        //move back to the camera for next picture
-                        self.moveAction()
-                    }
-                })
-            }
-        }
-        
-        DispatchQueue.main.async {
-            //update progress bar
-            uploadTask.observe(.progress) { [weak self] (snapshot) in
-                guard let strongSelf = self else {return}
-                guard let progress = snapshot.progress else { return}
-                strongSelf.progressView.progress = Float(progress.fractionCompleted)
-            }
         }
     }
